refactor: 컴포넌트에서 훅 분리
