+++ service-apply/src/components/home/HomeLogin.tsx
@@ -1,58 +1,17 @@
 import { Button, InputText, Txt } from '@quokka/design-system';
-import { useState } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
-import { useLoginMutate } from '../../hooks/react-query/useUser';
-import { isEmail, isPassword } from '../../functions/validator';
+import { Link } from 'react-router-dom';
+import useLoginForm from '../../hooks/home/useLoginForm';
 
 export const HomeLogin = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [isError, setIsError] = useState(false);
-  const [errorMessage, setErrorMessage] = useState('');
-
-  const { postLogin } = useLoginMutate();
-  const navigate = useNavigate();
-
-  const formAction = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (!email) {
-      alert('이메일을 입력해 주세요.');
-      return;
-    }
-    if (!password) {
-      alert('비밀번호를 입력해 주세요.');
-      return;
-    }
-    if (!isEmail(email)) {
-      alert('올바른 형식의 이메일을 입력해 주세요.');
-      setIsError(true);
-      setErrorMessage('올바른 형식의 이메일을 입력해 주세요.');
-      return;
-    }
-    if (!isPassword(password)) {
-      alert(
-        '비밀번호는 최소 8자 이상, 16자리 이하이며 최소 하나의 영문자, 숫자, 특수문자(!@#$%^&*)가 포함되어야 합니다.',
-      );
-      setIsError(true);
-      setErrorMessage(
-        '비밀번호는 최소 8자 이상, 16자리 이하이며 최소 하나의 영문자, 숫자, 특수문자(!@#$%^&*)가 포함되어야 합니다.',
-      );
-      return;
-    }
-    postLogin(
-      { email, pwd: password },
-      {
-        onError: (error) => {
-          alert(error.message);
-          setIsError(true);
-          setErrorMessage(error.message);
-        },
-        onSuccess: () => {
-          navigate('/apply');
-        },
-      },
-    );
-  };
+  const {
+    formAction,
+    email,
+    onChangeEmail,
+    password,
+    onChangePassword,
+    isError,
+    errorMessage,
+  } = useLoginForm();
 
   return (
     <div className="flex justify-end max-sm:mb-4">
@@ -69,10 +28,7 @@ export const HomeLogin = () => {
               name="email"
               className="w-full p-4"
               value={email}
-              onChange={(e) => {
-                setEmail(e.target.value);
-                setIsError(false);
-              }}
+              onChange={onChangeEmail}
             />
           </div>
           <div className="w-full">
@@ -83,10 +39,7 @@ export const HomeLogin = () => {
               name="password"
               className="w-full p-4"
               value={password}
-              onChange={(e) => {
-                setPassword(e.target.value);
-                setIsError(false);
-              }}
+              onChange={onChangePassword}
             />
           </div>
           <Link to={'/password-reset'}>
